Add files via upload
Coronavirus (COVID-19) is an illness caused by a virus that has spread throughout the world. COVID-19 symptoms can range from mild (or no symptoms) to severe illness. Since the first identified case in December 2019, the virus has resulted in an ongoing pandemic and causes more than 14.7 million cases across 188 countries and territories. The US has one of the most serious coronavirus outbreaks. According to Johns Hopkins University, there are over 380 million confirmed cases and over 141k deaths in the U.S. Covid19 also exerted negative impacts on the economy. The unemployment rate is an important indicator that shows the consequences of the pandemic.